Ensure transient consumer is disposed when the queue is deleted
Previously, the underlying factory would simply remove it from its list,
upon the receipt of 'basic.cancel' from the server.  Unfortunately, it
would not request for resources to be cleaned up and would end up leaving
consumers (and channels by extension) open indefinitely until the
underlying connection was lost.

It is yet to be determined if this should apply to PersistentConsumer, but
since it does very little when cancel occurs, it doesn't seem to apply.