add SerializeFriendly to TypeNameSerializer
Generic message types contain the assembly qualified type names for the
generic arguments.  This was causing an issue in which a new exchange
and queue would be created with each build due to the version numbers
changing in these type names.

IFriendlyTypeNameSerializer is an effort to expose a more user friendly
naming convention for the exchanges and queues that should not have the
issue as mentioned above.