BUGFIX: IBus.Subscribe[Async] method throws an error when the queue is previously created on RabbitMQ with different args than the default
When subscribing to an exchange using the IBus.Subscribe[Async] method
both queue and exchange are declared. If the queue was previously
created on RabbitMQ using arguments different than the default values on
IAdvancedBus.QueueDeclare[Async] method, the RabbitMQ will thrown the
following error:
The AMQP operation was interrupted: AMQP close-reason, initiated by
Peer, code=406, text="PRECONDITION_FAILED - inequivalent arg
'x-max-priority' for queue 'my_queue' in vhost '/': received none but
current is the value '9' of type 'long' [...]

To solve it, the user should be able to inject the arguments of the
queue declaration in the IBus.Subscribe[Async] method call, so the
IAdvancedBus.QueueDeclare[Async] declaration can match the previously
created queue. The existing argument Action<ISubscriptionConfiguration>
was used as inspiration to do the same changing the exchange and queue
args declarations.