Fix Client Command Dispatcher Thread hijacking.
`ClientCommandDispatcherSingleton.IvokeAsync` calls `Task.TrySetResultSafe` and sometimes continuations are executed synchronously as part of this call.  This is a [known trap](http://stackoverflow.com/q/22579206/458723) of the TPL.

TaskHelpers.TrySetResultSafe` tries to avoid hijacking of the thread by calling `Task.Run`.  However, `Task.Run` doesn't guarantee that the code will be executed on a different thread.  If `Task.Run` will execute continuations synchronously then Client Command Dispatcher Thread will execute code of the consumers.  If that code contains calls of `Publish` or `PublishAsync` or `QueueDeclare` and so on then the Client Command Dispatcher Thread will try to add an action to the queue inside `ClientCommandDispatcherSingleton`.  At the same time the Client Command Dispatcher Thread should consume this queue.  This is how we get a dead lock.