Make GetSupersededType work with any message version hierarchy (#902)
MessageVersionStack.GetSupersededType uses Type.GetInterfaces which does not
offer any order guarantees. When selecting the superseded type for a message, we could sort
by class name, but that would just introduce a limitation on message naming.

Instead of that, let's call GetInterfaces for type and type.BaseType, and grab the ISupersede
entries from those. The one that does not exist in base type's list is the correct superseded type.